refactor: redux toolkit을 적용한다
