+++ src/store/reducer.js
@@ -1,3 +1,7 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+import { fetchCategories, fetchRegions, fetchRestaurants } from '../services/api';
+
 const initialState = {
   regions: [],
   categories: [],
@@ -6,19 +10,34 @@ const initialState = {
   restaurants: [],
 };
 
-const reducers = {
-  UPDATE_FIELD: (state, action) => ({
-    ...state,
-    [action.payload.field]: action.payload.value,
-  }),
+const slice = createSlice({
+  name: 'field',
+  initialState,
+  reducers: {
+    updateField(state, action) {
+      return {
+        ...state,
+        [action.payload.field]: action.payload.value,
+      };
+    },
+  },
+});
+
+export const { updateField } = slice.actions;
+
+export const loadCategories = () => async (dispatch) => {
+  const categories = await fetchCategories();
+  dispatch(updateField({ field: 'categories', value: categories }));
 };
 
-const reducer = (state = initialState, action) => {
-  try {
-    return reducers[action.type](state, action);
-  } catch (error) {
-    return state;
-  }
+export const loadRegions = () => async (dispatch) => {
+  const regions = await fetchRegions();
+  dispatch(updateField({ field: 'regions', value: regions }));
 };
 
-export default reducer;
+export const loadRestaurants = ({ regionName, categoryId }) => async (dispatch) => {
+  const restaurants = await fetchRestaurants({ regionName, categoryId });
+  dispatch(updateField({ field: 'restaurants', value: restaurants }));
+};
+
+export default slice.reducer;
